% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_category.R
\name{create_category}
\alias{create_category}
\title{Create a New Category}
\usage{
create_category(
  store_id,
  name,
  description = NULL,
  handle = NULL,
  parent = NULL,
  subcategories = NULL,
  google_shopping_category = NULL,
  access_token = Sys.getenv("TIENDANUBE_ACCESS_TOKEN")
)
}
\arguments{
\item{store_id}{The ID of the store.}

\item{name}{List of the names of the Category, in every language supported by the store.}

\item{description}{List of the descriptions of the Category, as HTML, in every language supported
by the store.}

\item{handle}{List of the url-friendly strings generated from the Category's names, in every
language supported by the store.}

\item{parent}{Id of the Category's parent.}

\item{subcategories}{The ids of the Category's first level subcategories.}

\item{google_shopping_category}{Attributes used to categorize an item. This category is selected
from the Google’s taxonomy. The full list of product categories can be found here:
https://www.google.com/basepages/producttype/taxonomy.es-ES.txt .}

\item{access_token}{The store's access token for your app.}
}
\description{
Needs "Edit Products" permissions.
Docs: https://tiendanube.github.io/api-documentation/resources/category .
}
