% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_customer.R
\name{create_customer}
\alias{create_customer}
\title{Create a New Customer}
\usage{
create_customer(
  store_id,
  name,
  email = NULL,
  phone = NULL,
  identification = NULL,
  note = NULL,
  default_address = NULL,
  addresses = NULL,
  billing_address = NULL,
  billing_number = NULL,
  billing_floor = NULL,
  billing_locality = NULL,
  billing_zipcode = NULL,
  billing_city = NULL,
  billing_province = NULL,
  billing_country = NULL,
  extra = NULL,
  total_spent = NULL,
  total_spent_currency = NULL,
  last_order_id = NULL,
  active = TRUE,
  accepts_marketing = TRUE,
  send_email_invite = FALSE,
  password = NULL,
  access_token = Sys.getenv("TIENDANUBE_ACCESS_TOKEN")
)
}
\arguments{
\item{store_id}{The ID of the store.}

\item{name}{Name of the Customer.}

\item{email}{E-mail of the Customer.}

\item{phone}{Phone number of the customer (not necessarily the same as the address's phone).}

\item{identification}{Customer's identification (in Brazil for example, it would be the
CPF/CNPJ).}

\item{note}{Store owner's notes about the customer.}

\item{default_address}{Default shipping address of the Customer.}

\item{addresses}{List of shipping addresses for the Customer.}

\item{billing_address}{Billing address of the Customer.}

\item{billing_number}{Billing number of the Customer.}

\item{billing_floor}{Billing floor of the Customer.}

\item{billing_locality}{Billing locality of the Customer.}

\item{billing_zipcode}{Billing zipcode of the Customer.}

\item{billing_city}{Billing city of the Customer.}

\item{billing_province}{Billing province of the Customer.}

\item{billing_country}{Billing country code of the Customer.}

\item{extra}{A JSON object containing custom information. Can be set via the API or through
custom form fields of name "extra[key]" on the Customer's register form in the storefront.}

\item{total_spent}{The total amount of money that the Customer has spent at the store.}

\item{total_spent_currency}{The total spent's currency in ISO 4217 format.}

\item{last_order_id}{The id of the Customer's last Order.}

\item{active}{`TRUE` if the Customer activated his account. `FALSE` if he/she hasn't.}

\item{accepts_marketing}{Boolean field. Indicates if the buyer accepted to receive offers and
news via email. Read-only field in the API.}

\item{send_email_invite}{Send an email to notify the customer of their registration.}

\item{password}{User's password description.}

\item{access_token}{The store's access token for your app.}
}
\description{
Needs "Edit Customers" permissions.
Docs: https://tiendanube.github.io/api-documentation/resources/customer .
}
