% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_order.R
\name{create_order}
\alias{create_order}
\title{Create a New Order}
\usage{
create_order(
  store_id,
  products,
  customer,
  currency = NULL,
  language = NULL,
  gateway = NULL,
  payment_status = NULL,
  status = NULL,
  fulfillment = NULL,
  total = NULL,
  inventory_behaviour = NULL,
  note = NULL,
  billing_address = NULL,
  shipping_address,
  shipping_pickup_type = NULL,
  shipping = NULL,
  shipping_option = NULL,
  shipping_tracking_number = NULL,
  shipping_cost_customer = NULL,
  shipping_cost_owner = NULL,
  send_confirmation_email = FALSE,
  send_fulfillment_email = FALSE,
  location_id = NULL,
  access_token = Sys.getenv("TIENDANUBE_ACCESS_TOKEN")
)
}
\arguments{
\item{store_id}{The ID of the store.}

\item{products}{Order's products list (Product).}

\item{customer}{The customer object (Customer).}

\item{currency}{The order currency code (ISO 4217 format).}

\item{language}{The language code (ISO 639-1 format).}

\item{gateway}{The order's payment gateway (Payment Gateway).}

\item{payment_status}{The order's payment status (Payment Status).}

\item{status}{The order's status (Order Status).}

\item{fulfillment}{The order's status (Order Status).}

\item{total}{The sum of all products prices, shipping costs and discounts. Must be positive. If
not specified, it's calculated considering the provided costs and discounts.}

\item{inventory_behaviour}{The inventory behaviur that the order must perform (Inventory
Behaviour).}

\item{note}{An additional customer note for the order.}

\item{billing_address}{The customer's billing address object (Address).}

\item{shipping_address}{The customer's shipping address object (Address).}

\item{shipping_pickup_type}{The shipping pickup type (Shipping Type).}

\item{shipping}{The shipping method (Shipping Method).}

\item{shipping_option}{The order's shipping option nice name.}

\item{shipping_tracking_number}{The order's shipping tracking number}

\item{shipping_cost_customer}{The customer's shipping cost double value. The value 0 means free
shipping.}

\item{shipping_cost_owner}{The owner's shipping cost double value.}

\item{send_confirmation_email}{Send the order confirmation email to the customer.}

\item{send_fulfillment_email}{Send the order fulfillment email to the customer.}

\item{location_id}{Location id from where the stock will be decreased. Must be string.}

\item{access_token}{The store's access token for your app.}
}
\description{
Needs "Edit Orders" permissions.
Docs: https://tiendanube.github.io/api-documentation/resources/order .
}
