% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_product.R
\name{create_product}
\alias{create_product}
\title{Create a new Product}
\usage{
create_product(
  store_id,
  name,
  description = NULL,
  handle = NULL,
  variants,
  images = NULL,
  categories = NULL,
  brand = NULL,
  published = TRUE,
  free_shipping = FALSE,
  video_url = NULL,
  seo_title = NULL,
  seo_description = NULL,
  attributes = NULL,
  tags = NULL,
  requires_shipping = TRUE,
  access_token = Sys.getenv("TIENDANUBE_ACCESS_TOKEN")
)
}
\arguments{
\item{store_id}{The ID of the store.}

\item{name}{List of the names of the Product, in every language supported by the store.}

\item{description}{List of the descriptions of the Product, as HTML, in every language supported
by the store.}

\item{handle}{List of the url-friendly strings generated from the Product's names, in every
language supported by the store.}

\item{variants}{List of Product Variant objects representing the different version of the
Product.}

\item{images}{List of Product Image objects representing the Product's images.}

\item{categories}{List of Category Ids representing the Product's categories.}

\item{brand}{The Product's brand.}

\item{published}{`TRUE` if the Product is published in the store. `FALSE` otherwise.}

\item{free_shipping}{`TRUE` if the Product is eligible for free shipping. `FALSE` otherwise.}

\item{video_url}{String with a valid URL format. Only admits https links.}

\item{seo_title}{The SEO friendly title for the Product. Up to 70 characters.}

\item{seo_description}{The SEO friendly description for the Product. Up to 320 characters.}

\item{attributes}{List of the names of the attributes whose values define the variants.
E.g.: Color, Size, etc. It is important that the number of attributes is equal to the number
of values within the variants. Each product can have a maximum of 3 attributes.}

\item{tags}{String with all the Product's tags, separated by commas.}

\item{requires_shipping}{`TRUE` if the Product is physical. `FALSE` if it is digital description.}

\item{access_token}{The store's access token for your app.}
}
\description{
Needs "Edit Products" permissions.
Docs: https://tiendanube.github.io/api-documentation/resources/product .
}
