% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abandoned_checkouts.R
\name{get_abandoned_checkouts}
\alias{get_abandoned_checkouts}
\title{Get Abandoned Checkouts}
\usage{
get_abandoned_checkouts(
  store_id,
  page,
  created_at_max = NULL,
  updated_at_max = NULL,
  per_page = 200,
  fields = NULL,
  access_token = Sys.getenv("TIENDANUBE_ACCESS_TOKEN")
)
}
\arguments{
\item{store_id}{The ID of the store.}

\item{page}{Page to retrieve.}

\item{created_at_max}{Show Abandoned Checkouts created before date (ISO 8601 format).}

\item{updated_at_max}{Show Abandoned Checkouts last updated before date (ISO 8601 format).}

\item{per_page}{Amount of results.}

\item{fields}{Comma-separated list of fields to include in the response.}

\item{access_token}{The store's access token for your app.}
}
\description{
Receive a list of Abandoned Checkouts.
Docs: https://tiendanube.github.io/api-documentation/resources/abandoned-checkout .
}
