% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_access_token.R
\name{get_access_token}
\alias{get_access_token}
\title{Get Access Token}
\usage{
get_access_token(client_id, client_secret, code)
}
\arguments{
\item{client_id}{The Client or App Tienda Nube ID.}

\item{client_secret}{The Client Tienda Nube Secret.}

\item{code}{The customer activation code.}
}
\description{
When a customer installs our app, it redirects to the "URL to redirect after installation" and
it adds a `code` value to the URL. E.g., if our redirect URL is https://myapp.com , after
app installation, the customer will be redirected to https://myapp.com?code=SOME_CODE .
}
