% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abandoned_checkouts.R
\name{get_all_abandoned_checkouts}
\alias{get_all_abandoned_checkouts}
\title{Get All Abandoned Checkouts}
\usage{
get_all_abandoned_checkouts(
  store_id,
  created_at_max = NULL,
  updated_at_max = NULL,
  access_token = Sys.getenv("TIENDANUBE_ACCESS_TOKEN")
)
}
\arguments{
\item{store_id}{The ID of the store.}

\item{created_at_max}{Show Abandoned Checkouts created before date (ISO 8601 format).}

\item{updated_at_max}{Show Abandoned Checkouts last updated before date (ISO 8601 format).}

\item{access_token}{The store's access token for your app.}
}
\description{
Receive a list of all Abandoned Checkouts.
}
