% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_categories.R
\name{get_all_categories}
\alias{get_all_categories}
\title{Get All Categories}
\usage{
get_all_categories(
  store_id,
  created_at_min = NULL,
  created_at_max = NULL,
  updated_at_min = NULL,
  updated_at_max = NULL,
  fields = NULL,
  max_pages = Inf,
  access_token = Sys.getenv("TIENDANUBE_ACCESS_TOKEN")
)
}
\arguments{
\item{store_id}{The ID of the store.}

\item{created_at_min}{Show Categories created after date (ISO 8601 format).}

\item{created_at_max}{Show Categories created before date (ISO 8601 format).}

\item{updated_at_min}{Show Categories last updated after date (ISO 8601 format).}

\item{updated_at_max}{Show Categories last updated before date (ISO 8601 format).}

\item{fields}{Comma-separated list of fields to include in the response.}

\item{max_pages}{Max pages to query.}

\item{access_token}{The store's access token for your app.}
}
\description{
Receive a list of all Categories.
}
