% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_customers.R
\name{get_all_customers}
\alias{get_all_customers}
\title{Get All Customers}
\usage{
get_all_customers(
  store_id,
  created_at_min = NULL,
  created_at_max = NULL,
  updated_at_min = NULL,
  updated_at_max = NULL,
  access_token = Sys.getenv("TIENDANUBE_ACCESS_TOKEN")
)
}
\arguments{
\item{store_id}{The ID of the store.}

\item{created_at_min}{Show Customers created after date (ISO 8601 format).}

\item{created_at_max}{Show Customers created before date (ISO 8601 format).}

\item{updated_at_min}{Show Customers last updated after date (ISO 8601 format).}

\item{updated_at_max}{Show Customers last updated before date (ISO 8601 format).}

\item{access_token}{The store's access token for your app.}
}
\description{
Receive a list of all Customers.
}
