% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orders.R
\name{get_all_orders}
\alias{get_all_orders}
\title{Get All Orders}
\usage{
get_all_orders(
  store_id,
  created_at_min = NULL,
  created_at_max = NULL,
  updated_at_min = NULL,
  updated_at_max = NULL,
  access_token = Sys.getenv("TIENDANUBE_ACCESS_TOKEN")
)
}
\arguments{
\item{store_id}{The ID of the store.}

\item{created_at_min}{Show Orders created after date (ISO 8601 format).}

\item{created_at_max}{Show Orders created before date (ISO 8601 format).}

\item{updated_at_min}{Show Orders last updated after date (ISO 8601 format).}

\item{updated_at_max}{Show Orders last updated before date (ISO 8601 format).}

\item{access_token}{The store's access token for your app.}
}
\description{
Receive a list of all Orders.
}
