% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_products.R
\name{get_all_products}
\alias{get_all_products}
\title{Get All Products}
\usage{
get_all_products(
  store_id,
  access_token = Sys.getenv("TIENDANUBE_ACCESS_TOKEN")
)
}
\arguments{
\item{store_id}{The ID of the store.}

\item{access_token}{The store's access token for your app.}
}
\description{
Receive a list of all Products.
}
