% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_customers.R
\name{get_customers}
\alias{get_customers}
\title{Get Customers}
\usage{
get_customers(
  store_id,
  page,
  created_at_min = NULL,
  created_at_max = NULL,
  updated_at_min = NULL,
  updated_at_max = NULL,
  per_page = 200,
  fields = NULL,
  access_token = Sys.getenv("TIENDANUBE_ACCESS_TOKEN")
)
}
\arguments{
\item{store_id}{The ID of the store.}

\item{page}{Page to retrieve.}

\item{created_at_min}{Show Customers created after date (ISO 8601 format).}

\item{created_at_max}{Show Customers created before date (ISO 8601 format).}

\item{updated_at_min}{Show Customers last updated after date (ISO 8601 format).}

\item{updated_at_max}{Show Customers last updated before date (ISO 8601 format).}

\item{per_page}{Amount of results.}

\item{fields}{Comma-separated list of fields to include in the response.}

\item{access_token}{The store's access token for your app.}
}
\description{
Receive a list of Customers.
Needs "View Customers" permissions.
Docs: https://tiendanube.github.io/api-documentation/resources/customer .
}
