% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_products.R
\name{get_products}
\alias{get_products}
\title{Get Products}
\usage{
get_products(
  store_id,
  page,
  per_page = 200,
  access_token = Sys.getenv("TIENDANUBE_ACCESS_TOKEN")
)
}
\arguments{
\item{store_id}{The ID of the store.}

\item{page}{Page to retrieve.}

\item{per_page}{Amount of results.}

\item{access_token}{The store's access token for your app.}
}
\description{
Receive a list of Products.
Docs: https://tiendanube.github.io/api-documentation/resources/product .
}
