% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_store.R
\name{get_store}
\alias{get_store}
\title{Get Store}
\usage{
get_store(store_id, access_token = Sys.getenv("TIENDANUBE_ACCESS_TOKEN"))
}
\arguments{
\item{store_id}{The ID of the store.}

\item{access_token}{The store's access token for your app.}
}
\description{
The Store resource contains general settings and information about a Tiendanube's store.
Docs: https://tiendanube.github.io/api-documentation/resources/store .
}
