% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiendanube-package.R
\docType{package}
\name{tiendanube-package}
\alias{tiendanube}
\alias{tiendanube-package}
\title{Tienda Nube API Interface.}
\description{
Tienda Nube API Interface. The official API documentaion is at \url{https://tiendanube.github.io/api-documentation}.
}
\author{
\strong{Maintainer}: Juan Cruz Rodriguez \email{jcrodriguez@unc.edu.ar}

}
