% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiendanube-package.R
\docType{package}
\name{tiendanube-package}
\alias{tiendanube}
\alias{tiendanube-package}
\title{Tienda Nube API interface.}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
